+++ AdvancedMapAndroid/app/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name" >
         <activity
-            android:name=".map_list.LauncherListActivity"
+            android:name=".maplist.LauncherListActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -26,7 +26,7 @@
             </intent-filter>
         </activity>
         <activity android:name=".map.PinMapActivity" />
-        <activity android:name=".map_list.PackageManagerActivity" />
+        <activity android:name=".maplist.PackageManagerActivity" />
         <activity android:name=".map.PackagedMapActivity" />
         <activity android:name=".map.Overlays2DActivity" />
         <activity android:name=".map.MapListenerActivity" />
