+++ AdvancedMapAndroid/app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:label="@string/app_name"
         android:name=".AdvancedMapApplication">
         <activity
-            android:name=".map_list.LauncherListActivity"
+            android:name=".maplist.LauncherListActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -48,7 +48,7 @@
         <activity android:name=".map.MapZenActivity" />
         <activity android:name=".map.CaptureActivity" />
 
-        <activity android:name=".map_list.PackageManagerActivity" />
+        <activity android:name=".maplist.PackageManagerActivity" />
 
         <activity android:name="com.carto.filepicker.FilePicker" />
