+++ AdvancedMapAndroid/app/src/main/java/com/carto/advancedmap/map/AnimatedRasterMapActivity.java
@@ -3,7 +3,7 @@ package com.carto.advancedmap.map;
 import android.os.Bundle;
 
 import com.carto.advancedmap.Description;
-import com.carto.advancedmap.map_base.VectorMapSampleBaseActivity;
+import com.carto.advancedmap.mapbase.VectorMapSampleBaseActivity;
 import com.carto.advancedmap.datasource.MyAnimatedTileDataSource;
 import com.carto.datasources.MemoryCacheTileDataSource;
 import com.carto.datasources.HTTPTileDataSource;
