+++ AdvancedMapAndroid/app/src/main/java/com/carto/advancedmap/map/CaptureActivity.java
@@ -7,7 +7,7 @@ import android.os.Environment;
 
 import com.carto.advancedmap.Description;
 import com.carto.advancedmap.R;
-import com.carto.advancedmap.map_base.VectorMapSampleBaseActivity;
+import com.carto.advancedmap.mapbase.VectorMapSampleBaseActivity;
 import com.carto.core.MapPos;
 import com.carto.core.MapRange;
 import com.carto.datasources.LocalVectorDataSource;
