+++ AdvancedMapAndroid/app/src/main/java/com/carto/advancedmap/map/CustomPopupActivity.java
@@ -6,7 +6,7 @@ import android.os.Bundle;
 
 import com.carto.advancedmap.Description;
 import com.carto.advancedmap.R;
-import com.carto.advancedmap.map_base.VectorMapSampleBaseActivity;
+import com.carto.advancedmap.mapbase.VectorMapSampleBaseActivity;
 import com.carto.advancedmap.vectorelements.MyCustomPopupHandler;
 import com.carto.core.MapPos;
 import com.carto.datasources.LocalVectorDataSource;
