+++ AdvancedMapAndroid/app/src/main/java/com/carto/advancedmap/map/CustomRasterDataSourceActivity.java
@@ -3,7 +3,7 @@ package com.carto.advancedmap.map;
 import android.os.Bundle;
 
 import com.carto.advancedmap.Description;
-import com.carto.advancedmap.map_base.MapSampleBaseActivity;
+import com.carto.advancedmap.mapbase.MapSampleBaseActivity;
 import com.carto.advancedmap.datasource.MyMergedRasterTileDataSource;
 import com.carto.advancedmap.util.Const;
 import com.carto.core.MapPos;
