+++ AdvancedMapAndroid/app/src/main/java/com/carto/advancedmap/map/OfflineRoutingActivity.java
@@ -12,7 +12,7 @@ import android.widget.Toast;
 
 import com.carto.advancedmap.Description;
 import com.carto.advancedmap.R;
-import com.carto.advancedmap.map_base.VectorMapSampleBaseActivity;
+import com.carto.advancedmap.mapbase.VectorMapSampleBaseActivity;
 import com.carto.advancedmap.util.Const;
 import com.carto.core.MapPos;
 import com.carto.core.MapPosVector;
