+++ AdvancedMapAndroid/app/src/main/java/com/carto/advancedmap/map/OfflineVectorMapActivity.java
@@ -4,7 +4,7 @@ import android.os.Bundle;
 import android.util.Log;
 
 import com.carto.advancedmap.Description;
-import com.carto.advancedmap.map_base.VectorMapSampleBaseActivity;
+import com.carto.advancedmap.mapbase.VectorMapSampleBaseActivity;
 import com.carto.advancedmap.util.AssetCopy;
 import com.carto.advancedmap.util.Const;
 import com.carto.core.MapRange;
