+++ AdvancedMapAndroid/app/src/main/java/com/carto/advancedmap/mapbase/VectorMapSampleBaseActivity.java
@@ -1,4 +1,4 @@
-package com.carto.advancedmap.map_base;
+package com.carto.advancedmap.mapbase;
 import android.os.Bundle;
 
 import android.util.Log;
