+++ CartoMapAndroid/app/src/main/java/com/carto/cartomap/android/map/CartoUTFGridActivity.java
@@ -4,8 +4,8 @@ import android.os.Bundle;
 import android.util.Log;
 
 import com.carto.cartomap.android.Description;
-import com.carto.cartomap.android.map_base.VectorMapSampleBaseActivity;
-import com.carto.cartomap.android.map_listener.MyUTFGridEventListener;
+import com.carto.cartomap.android.mapbase.VectorMapSampleBaseActivity;
+import com.carto.cartomap.android.maplistener.MyUTFGridEventListener;
 import com.carto.core.MapPos;
 import com.carto.core.Variant;
 import com.carto.datasources.LocalVectorDataSource;
