+++ CartoMapAndroid/app/src/main/java/com/carto/cartomap/android/map/CartoVisJSONActivity.java
@@ -6,7 +6,7 @@ import android.view.Menu;
 import android.view.MenuItem;
 
 import com.carto.cartomap.android.Description;
-import com.carto.cartomap.android.map_base.MapSampleBaseActivity;
+import com.carto.cartomap.android.mapbase.MapSampleBaseActivity;
 import com.carto.core.MapPos;
 import com.carto.core.Variant;
 import com.carto.core.VariantType;
