+++ CartoMapAndroid/app/src/main/java/com/carto/cartomap/android/maplistener/MyUTFGridEventListener.java
@@ -1,4 +1,4 @@
-package com.carto.cartomap.android.map_listener;
+package com.carto.cartomap.android.maplistener;
 
 import android.util.Log;
